Update test name
