+++ Tests/ReferenceImages_64/StripePaymentSheetTests.EmbeddedViewRepresentableSnapshotTest/testEmbeddedPaymentElementViewWithCompletionHandler__after_height_change@3x.png
