+++ Tests/ReferenceImages_64/StripePaymentSheetTests.EmbeddedViewRepresentableSnapshotTest/testEmbeddedPaymentElementViewWithCompletionHandler__after_third_height_change@3x.png
