+++ Tests/ReferenceImages_64/StripePaymentSheetTests.EmbeddedViewRepresentableSnapshotTest/testEmbeddedPaymentElementViewWithCompletionHandler__before_height_change@3x.png
